Significant work including renaming generated types, fixing generic rpcs and separating format from transport

+++ macros_impl/src/tests/inputs/resource.rs
@@ -1,6 +1,6 @@
 #[rpc]
 pub trait Resources<T> {
     fn list(&self) -> Vec<T>;
-    fn get(&self, id: usize) -> Option<T>;
+    fn get(&self, id: u64) -> Option<T>;
     fn new(&self, value: T);
 }
